requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.1
click==8.0.1
idna==3.2
joblib==1.0.1
nltk==3.6.2
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
soupsieve==2.2.1
tqdm==4.61.2
urllib3==1.26.6
